lib_path in own file to help external reuse/clarity
