+++ src/CMakeLists.txt
@@ -6,14 +6,6 @@ target_compile_definitions(filesystem_cpp PRIVATE
 $<$<BOOL:${MSVC}>:_CRT_SECURE_NO_WARNINGS>
 )
 
-#--- optional lib_path()
-if(WIN32)
-  target_compile_definitions(filesystem_cpp PRIVATE FS_DLL_NAME="$<TARGET_FILE_NAME:filesystem>")
-elseif(HAVE_DLADDR)
-  target_link_libraries(filesystem_cpp PRIVATE ${CMAKE_DL_LIBS})
-  target_compile_definitions(filesystem_cpp PRIVATE _GNU_SOURCE HAVE_DLADDR)
-endif()
-
 # --- source files
 set(s ${CMAKE_CURRENT_SOURCE_DIR})
