+++ src/max.c
@@ -1,0 +1,3 @@
+#include "filesystem.h"
+
+size_t get_maxp(){ return MAXP; }
